+++ examples/hello.cpp
@@ -1,6 +1,7 @@
-#include <fluxgl/fluxgl.h>
 #include <iostream>
 #include <sstream>
+
+#include <fluxgl/fluxgl.h>
 
 int main() {
     try {
@@ -59,33 +60,7 @@ int main() {
         glDeleteShader(vertexShader);
         glDeleteShader(fragmentShader);
         
-        unsigned int VAO;
-        glGenVertexArrays(1, &VAO);
-        glBindVertexArray(VAO);
-
-        unsigned int VBO;
-        glGenBuffers(1, &VBO);
-        glBindBuffer(GL_ARRAY_BUFFER, VBO);
-        glBufferData(GL_ARRAY_BUFFER, sizeof(float) * 12, (float[]) {
-            -0.5f, -0.5f, 0.0f,
-             0.5f, -0.5f, 0.0f,
-             0.5f,  0.5f, 0.0f,
-            -0.5f,  0.5f, 0.0f
-        }, GL_STATIC_DRAW);
-
-        unsigned int EBO;
-        glGenBuffers(1, &EBO);
-        glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, EBO);
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(unsigned int) * 6, (unsigned int[]) {
-            0, 1, 2,
-            2, 3, 0
-        }, GL_STATIC_DRAW);
-
-        glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float), (void*)0);
-        glEnableVertexAttribArray(0);
-
-        glBindBuffer(GL_ARRAY_BUFFER, 0);
-        glBindVertexArray(0);
+        fluxgl::Mesh rectangle = fluxgl::Mesh::quad();
 
         while (!window.shouldClose()) {
             window.pollEvents();
@@ -94,8 +69,7 @@ int main() {
             glClear(GL_COLOR_BUFFER_BIT);
 
             glUseProgram(shaderProgram);
-            glBindVertexArray(VAO);
-            glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, 0);
+            rectangle.draw();
 
             window.swapBuffers();
         }