+++ include/fluxgl/fluxgl.h
@@ -3,3 +3,6 @@
 #include "core/window.h"
 #include "core/error.h"
 #include "core/log.h"
+
+#include "graphics/vertex.h"
+#include "graphics/mesh.h"