+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Minimal logging system with debug, info, warning, and error levels
+- Mesh class with vertex and index buffers
 
 ## [0.1.1] - 2026-02-11